python/requirements.txt
pymongo[srv]
python-dotenv
python-dateutil
